Fix device-install.bat baud rate  (master --> develop) (#7816)
* Upgrade trunk (#7763)

Co-authored-by: vidplace7 <1779290+vidplace7@users.noreply.github.com>

* Fix device-install.bat baud rate

As reported by @gruberaaron , work to improve the 1200bps reset for
esptool caused all runs of device-install.bat to use 1200bps as
the baud rate.

This change removes the general SET "ESPTOOL_BAUD=1200" that was causing
the issues and places the baud settings for reset mode inside the conditional.

Fixes https://github.com/meshtastic/firmware/issues/7172

---------

Co-authored-by: github-actions[bot] <41898282+github-actions[bot]@users.noreply.github.com>
Co-authored-by: vidplace7 <1779290+vidplace7@users.noreply.github.com>